refactor model.py into a subpackage
